requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.11
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
django-cors-headers==3.13.0
djangorestframework-simplejwt
